서버 작동 안되는 문제 해결 및 회원 동네 등록 api 완료

+++ src/main/java/ga/todayOutside/src/user/UserInfoRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository // => JPA => Hibernate => ORM => Database 객체지향으로 접근하게 해주는 도구이다
-public interface UserInfoRepository extends CrudRepository<UserInfo, Integer> {
+public interface UserInfoRepository extends CrudRepository<UserInfo, Long> {
     List<UserInfo> findByStatus(String status);
     Optional<UserInfo> findByEmail(String email);
     List<UserInfo> findByEmailAndStatus(String email, String status);
